Fix onRealloc code to trace reallocation only if it isn't already traced
Trace reallocation only if it isn't already traced by separate onFree and onMalloc calls

Sometimes reallocation involves calling malloc and free, which
would be traced in separate onMalloc and onFree calls, thus this case
should be ignored by onRealloc. We ensure this by checking pointer size,
which would be NaN if onFree is already called on it.